# restFul/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service center API",
        "title": "Service Center API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/admin/telegram-trust": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список всех доверенных чатов Telegram",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram-trust"
                ],
                "summary": "Получить список доверенных чатов Telegram",
                "operationId": "get-telegram-trust",
                "responses": {
                    "200": {
                        "description": "Массив информации о доверенных чатах Telegram",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.TrustedTelegramUsers"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать новую запись о доверенном чате Telegram",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram-trust"
                ],
                "summary": "Добавить доверенный чат Telegram",
                "operationId": "add-telegram-trust",
                "parameters": [
                    {
                        "description": "Информация о доверенном чате Telegram",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.TrustedTelegramUsers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    }
                }
            }
        },
        "/api/admin/telegram-trust/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновить информацию о доверенном чате Telegram",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram-trust"
                ],
                "summary": "Обновить информацию о доверенном чате Telegram",
                "operationId": "update-telegram-trust",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID доверенного чата Telegram",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация о доверенном чате Telegram",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.TrustedTelegramUsers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удалить информацию о доверенном чате Telegram",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram-trust"
                ],
                "summary": "Удалить доверенный чат Telegram",
                "operationId": "delete-telegram-trust",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID доверенного чата Telegram",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    }
                }
            }
        },
        "/api/task": {
            "post": {
                "description": "Создать новую задачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Добавить задачу",
                "operationId": "add-task",
                "parameters": [
                    {
                        "description": "Информация о заявке",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.ApplicationForCall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    }
                }
            }
        },
...
        "/auth/admin/sign-in": {
            "post": {
                "description": "Создает аккаунт",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация в CMS",
                "operationId": "create-account",
                "parameters": [
                    {
                        "description": "Информация об аккаунте",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.ApplicationForCall": {
            "type": "object",
            "required": [
                "name",
                "phoneNumber"
            ],
            "properties": {
                "date": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
...
        "database.TrustedTelegramUsers": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "telegramID": {
                    "type": "string"
                }
            }
        },
        "database.User": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Author",
            "in": "header"
        }
    }
}